Create Lecture: The Red Hat Package Manager (rpm)
Learn how to use the rpm command to manage packages on Red Hat based systems. You will see how to query, install, remove and upgrade software with the venerable rpm command. Also, we show you how to use the rpm2cpio command to extract the contents of an rpm package