Create Lecture: Your Bash Shell Enviroment
We dive into the Bash shell environment, and learn how to create our own Bash variables and functions. Not only is this helpful for now, but it will come in handy when we start discussing shell scripts. We also take a look at a few new commands, and find out how to modify our Bash configuration.