Update Lecture: Text File Statistics
This lesson explores how we can glean some information from our files, such as how many lines they contain, the amount of words within a file, and byte counts. We will use the nl, wc, and od commands to dive deeper into the contents and structure of text files. We also learn how to use hash values to verify the integrity of files.