Create Lecture: More Text Manipulation
In this lesson we learn some basic usage of the sed command. We also discuss the split command, and how you can put files back together after you broke them apart with split. Understanding these commands, as with all Bash commands that we discuss, are crucial for passing the LPIC-1 exam.