Update Lecture: Understanding Standard Input, Output, and Error
In this lesson, we work with the fundamentals of standard input, standard output, and standard error. We have already been using some of the concepts in the course up to this point, but now we are going to take a step back and find out what is going on 'under the hood.' We work with redirecting these messages to other commands and files, and what their specific file handles are.