Create Lecture: Reviewing the State of Your System
This lesson introduces you to how Linux processes work. We will utilize the ps and top commands to view running processes on a system and find out how we can filter the results from these commands. Every Linux system administrator needs to know what is going on with their systems in order to effectively manage them.