Create Lecture: Keep a process running
This lesson focuses on how to keep processes running on a system, even if you are not actively on the computer. We learn about the screen and tmux multiplexing window managers, which allow you to keep remote sessions active on a system. We also look at running jobs in the background, and how to keep tabs on those processes and stop them when you need to.