Create Lecture: Legacy MBR Partitions
This lesson will show you how to use the fdisk and parted commands to create legacy MBR partition tables. Note that a local virtual machine was used for this so we would have extra disks to create partitions on. It is recommended that you set up your own local virtual machine to try these commands out on your own.